Tests/LibWeb/Text/input/CSSOMView/mouse-event-offset-values.html

<!DOCTYPE html>
<style>
    div {
        width: 100px;
        height: 100px;
        background-color: gray;
        transform: rotate(45deg);
    }
</style>
<div onClick="
    println(`offsetX: ${event.offsetX}`);
    println(`offsetY: ${event.offsetY}`);
"></div>

<script src="../include.js"></script>

<script>
    test(() => {
        internals.click(50, 60);
    });
</script>
