Tests/LibWeb/Text/expected/CSSOMView/mouse-event-offset-values.txt

offsetX: 45.75
offsetY: 57.078125
